[DESCRIPTION]

Using clang-4.0 to compile our library I've noticed one warning about the promotion of types for the va_args arguments is undefined. This is the solution to that problem.

+++ include_senml/xively_senml.h
@@ -88,7 +88,7 @@ extern xi_state_t xi_senml_serialize( xi_senml_t* senml_structure,
  * @retval XI_STATE_OK if succeeded, other in case of failure,
  *         see xively_error.h for error codes
  */
-extern xi_state_t xi_create_senml_struct( xi_senml_t** senml_ptr, uint8_t count, ... );
+extern xi_state_t xi_create_senml_struct( xi_senml_t** senml_ptr, int count, ... );
 
 /**
  * @name xi_add_senml_entry
@@ -106,7 +106,7 @@ extern xi_state_t xi_create_senml_struct( xi_senml_t** senml_ptr, uint8_t count,
  * @retval XI_STATE_OK if succeeded, other in case of failure,
  *         see xively_error.h for error codes
  */
-extern xi_state_t xi_add_senml_entry( xi_senml_t* senml_ptr, uint8_t count, ... );
+extern xi_state_t xi_add_senml_entry( xi_senml_t* senml_ptr, int count, ... );
 
 /**
  * @name xi_senml_free_buffer
